Se acomodan los archivos de mejor forma
