+++ Gestor_De_Libros/archives/Libros/Caperucita Roja.txt
@@ -1,0 +1,9 @@
+Nombre del libro: Caperucita Roja
+
+Autor: Wilhelm i Jacob Grimm 
+
+Año de impresión: 1697, 1812
+
+ISBN: 978-84-7864-851-1
+
+No. Páginas: 24
