+++ Gestor_De_Libros/archives/Libros/La Biblia de CSharp - Anaya.txt
@@ -1,0 +1,9 @@
+Nombre del libro: La Biblia de CSharp - Anaya
+
+Autor: Jason Jeff Brian Beres
+
+Año de impresión: 2003
+
+ISBN: 978-844-151-484-3
+
+No. Páginas: 861
