+++ Gestor_De_Libros/archives/Libros/Los 3 Cochinitos.txt
@@ -1,0 +1,9 @@
+Nombre del libro: Los Tres Cochinitos
+
+Autor: Roberto Bravo Meritxell Marti
+
+Año de impresión: 1900
+
+ISBN: 978-84-7864-765-1
+
+No. Páginas: 24
