--- gestor-de-libros-master/Gestor_De_Libros.sln
@@ -1,18 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-# SharpDevelop 5.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Gestor_De_Libros", "Gestor_De_Libros\Gestor_De_Libros.csproj", "{5C1BFC04-0AAF-488A-9D94-CF4B7E3E90CA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5C1BFC04-0AAF-488A-9D94-CF4B7E3E90CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5C1BFC04-0AAF-488A-9D94-CF4B7E3E90CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5C1BFC04-0AAF-488A-9D94-CF4B7E3E90CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5C1BFC04-0AAF-488A-9D94-CF4B7E3E90CA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
