--- gestor-de-libros-master/Gestor_De_Libros/bin/Debug/Gestor_De_Libros.exe.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-	<startup>
-		<supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0" />
-	</startup>
-</configuration>
