--- gestor-de-libros-master/README.md
@@ -1,1 +1,0 @@
-# gestor-de-libros
